# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ai Pop Cute Support Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
  /users/by-username/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Get user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
  /users/me/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/privacy:
    patch:
      operationId: updatePrivacy
      tags: [users]
      summary: Update privacy settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivacyUpdate"
      responses:
        "200":
          description: Privacy settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacySettings"
  /users/me/notification-settings:
    get:
      operationId: getNotificationSettings
      tags: [users]
      summary: Get notification settings
      responses:
        "200":
          description: Notification settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
    patch:
      operationId: updateNotificationSettings
      tags: [users]
      summary: Update notification settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
  /users/me/chat-settings:
    patch:
      operationId: updateChatSettings
      tags: [users]
      summary: Update chat settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSettingsUpdate"
      responses:
        "200":
          description: Chat settings updated
  /users/{userId}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"
  /users/{userId}/unfollow:
    post:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfollowed
  /users/{userId}/followers:
    get:
      operationId: getFollowers
      tags: [users]
      summary: Get followers list
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Followers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}/following:
    get:
      operationId: getFollowing
      tags: [users]
      summary: Get following list
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Following list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get user's posts
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
  /users/{userId}/block:
    post:
      operationId: blockUser
      tags: [users]
      summary: Block a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User blocked
  /users/{userId}/unblock:
    post:
      operationId: unblockUser
      tags: [users]
      summary: Unblock a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unblocked
...
  /users/{userId}/restrict:
    post:
      operationId: restrictUser
      tags: [users]
      summary: Restrict a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User restricted
  /users/{userId}/unrestrict:
    post:
      operationId: unrestrictUser
      tags: [users]
      summary: Unrestrict a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unrestricted
...
  /users/follow-requests/{requesterId}/accept:
    post:
      operationId: acceptFollowRequest
      tags: [users]
      summary: Accept a follow request
      parameters:
        - name: requesterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request accepted
  /users/follow-requests/{requesterId}/reject:
    post:
      operationId: rejectFollowRequest
      tags: [users]
      summary: Reject a follow request
      parameters:
        - name: requesterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request rejected
  /users/me/activity-log:
    get:
      operationId: getActivityLog
      tags: [users]
      summary: Get user activity log
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Activity log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLogResponse"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, numericId, name, username, email, isVerified, role, createdAt]
      properties:
        id:
          type: integer
        numericId:
          type: integer
        name:
          type: string
        username:
          type: string
        email:
          type: string
        bio:
          type: ["string", "null"]
        alternateBio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        role:
          type: string
          enum: [user, creator]
        isPrivate:
          type: boolean
        mixChannelId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, numericId, name, username, isVerified, role, followersCount, followingCount, postsCount, likesCount, createdAt]
      properties:
        id:
          type: integer
        numericId:
          type: integer
        name:
          type: string
        username:
          type: string
        bio:
          type: ["string", "null"]
        alternateBio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        role:
          type: string
          enum: [user, creator]
        isPrivate:
          type: boolean
        mixChannelId:
          type: ["string", "null"]
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        likesCount:
          type: integer
        totalVotes:
          type: integer
        profileViews:
          type: integer
        isFollowing:
          type: ["boolean", "null"]
        isFollowRequestSent:
          type: ["boolean", "null"]
        isBlocked:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        bio:
          type: ["string", "null"]
        alternateBio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        mixChannelId:
          type: ["string", "null"]
    PrivacyUpdate:
      type: object
      properties:
        isPrivate:
          type: boolean
        whoCanMessage:
          type: string
          enum: [everyone, followers, none]
        whoCanComment:
          type: string
          enum: [everyone, followers, none]
        whoCanViewFollowers:
          type: string
          enum: [everyone, followers, none]
        whoCanDownload:
          type: string
          enum: [everyone, followers, none]
    PrivacySettings:
      type: object
      required: [isPrivate, whoCanMessage, whoCanComment, whoCanViewFollowers, whoCanDownload]
      properties:
        isPrivate:
          type: boolean
        whoCanMessage:
          type: string
        whoCanComment:
          type: string
        whoCanViewFollowers:
          type: string
        whoCanDownload:
          type: string
    NotificationSettings:
      type: object
      properties:
        votes:
          type: boolean
        newFollowers:
          type: boolean
        followRequests:
          type: boolean
        likes:
          type: boolean
        comments:
          type: boolean
        replies:
          type: boolean
        messages:
          type: boolean
        mentions:
          type: boolean
        postShares:
          type: boolean
        profileViews:
          type: boolean
        aiNewPosts:
          type: boolean
        systemAlerts:
          type: boolean
    NotificationSettingsUpdate:
      type: object
      properties:
        votes:
          type: boolean
        newFollowers:
          type: boolean
        followRequests:
          type: boolean
        likes:
          type: boolean
        comments:
          type: boolean
        replies:
          type: boolean
        messages:
          type: boolean
        mentions:
          type: boolean
        postShares:
          type: boolean
        profileViews:
          type: boolean
        aiNewPosts:
          type: boolean
        systemAlerts:
          type: boolean
    ChatSettingsUpdate:
      type: object
      properties:
        readReceipts:
          type: boolean
        typingIndicators:
          type: boolean
        showOnlineStatus:
          type: boolean
        whoCanMessageMe:
          type: string
          enum: [everyone, followers, none]
    FollowStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [following, requested, unfollowed]
    UserListResponse:
      type: object
      required: [users, total, page]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer

    # ── Post Schemas ─────────────────────────────────────────────────────────
    Post:
      type: object
      required: [id, authorId, visibility, allowComments, allowDownloads, isPinned, viewsCount, likesCount, commentsCount, sharesCount, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserProfile"
        caption:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
          enum: ["image", "video", null]
        filterName:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, followers, private]
        allowComments:
          type: boolean
        allowDownloads:
          type: boolean
        isPinned:
          type: boolean
        viewsCount:
          type: integer
        likesCount:
          type: integer
        commentsCount:
          type: integer
        sharesCount:
          type: integer
        downloadsCount:
          type: integer
        isLiked:
          type: ["boolean", "null"]
        isSaved:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PostListResponse:
      type: object
      required: [posts, total, page]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        hasMore:
          type: boolean
...
    ActivityLog:
      type: object
      required: [id, userId, action, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        action:
          type: string
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ActivityLogResponse:
      type: object
      required: [activities, total, page]
      properties:
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"
        total:
          type: integer
        page:
          type: integer
...