# lib/api-zod/src/generated/types/followStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ai Pop Cute Support Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FollowStatusStatus } from './followStatusStatus';

export interface FollowStatus {
  status: FollowStatusStatus;
}